some change: change NumericalField from int to double add field name to Product SellerController: add function(addProduct, createOptionalFields, createNumericalFields)

+++ src/main/java/model/product/Field/NumericalField.java
@@ -1,14 +1,14 @@
 package model.product.Field;
 
 public class NumericalField extends Field{
-    private int numericalField;
+    private double numericalField;
 
-    public NumericalField(String name, FieldType type, int numericalField) {
+    public NumericalField(String name, FieldType type, double numericalField) {
         super(name, type);
         this.numericalField = numericalField;
     }
 
-    public int getNumericalField() {
+    public double getNumericalField() {
         return numericalField;
     }
 }
